Create tempLoggerU6_14xLM35.py
A simple temperature logging script that logs the temperature of up to 14 LM35 temperature sensors that connect to the analog inputs of a U6 Labjack. It writes to the logging file in ~/rawData/tempLog/ about every second and the name of the logging file comprises the starting date/time of the measurement.